URLs for dashboard tabs (#816)
* Extract dashboard tabs to routes

* Fix dashboard tabs initial loading

* PowerUp and PowerDown containers, routes for powerup and powerdown

* Do not allow direct visit to invest page if it disabled
