+++ app/components/Dashboard/Tabs.js
@@ -1,16 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import Link from '../Link';
 
 import {
   Tab,
-  TabButton,
+  TabLink,
+  TabLinkActive,
   TabButtonDisabled,
   TabsWrapper,
   TabIcon,
   TabTitle,
 } from './styles';
 
-const Tabs = ({ activeTab, tabs, setActiveTab, disabledTabs }) => (
+const Tabs = ({ tabs, disabledTabs }) => (
   <TabsWrapper name="tabs">
     {tabs.map((tab) => {
       if (disabledTabs && disabledTabs.indexOf(tab.name) !== -1) {
@@ -23,15 +25,18 @@ const Tabs = ({ activeTab, tabs, setActiveTab, disabledTabs }) => (
           </Tab>
         );
       }
+
       return (
         <Tab name="tab" key={tab.name} data-tour={tab.name}>
-          <TabButton
-            disabled={tab.name === activeTab}
-            onClick={() => setActiveTab(tab.name)}
+          <Link
+            to={tab.to}
+            component={TabLink}
+            activeComponent={TabLinkActive}
+            onlyActiveOnIndex={tab.onlyActiveOnIndex}
           >
             <TabIcon className={`fa ${tab.icon}`} />
             <TabTitle>{tab.title}</TabTitle>
-          </TabButton>
+          </Link>
         </Tab>
       );
     }
@@ -39,10 +44,8 @@ const Tabs = ({ activeTab, tabs, setActiveTab, disabledTabs }) => (
   </TabsWrapper>
 );
 Tabs.propTypes = {
-  activeTab: PropTypes.string.isRequired,
   disabledTabs: PropTypes.array,
   tabs: PropTypes.array.isRequired,
-  setActiveTab: PropTypes.func.isRequired,
 };
 
 export default Tabs;
